requirements.txt
python-dotenv>=0.21.0
selenium>=4.4.3